Create Voting Election.c
This is a software that is built with voting elections in mind. Obviously there isn't enough memory for the computer to process too many votes, so this is just a passion project. I have taken an interest in elections specifically the software, so this is what I came up with.