Create Volume Changer.c
his program was built as many videos have low volume, especially for students doing online school. I wanted to be able to affect the volume of a video so a student would be able to hear the teacher better. This is similar to what your computer does when you click "increase" or "decrease" sound but you can multiply the sound.