Bloodtype.c
This is a program to determine what bloodtype a child will have. This is dependent on the parents blood types, and the blood types are A, B, and O. This is important as when receiving blood donations your body has to accept the type of blood  that the donated blood is.  This makes it easier for doctors to figure the blood  type needed by the patient.